requirements.txt
python-dotenv==1.2.1

# -------------------------
# RAG
# -------------------------
qdrant-client==1.10.1
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1
nltk==3.9.1
tiktoken==0.7.0
pydantic==2.9.2
python-dateutil==2.9.0.post0

# -------------------------
# LangChain + OpenAI (LLM)
# -------------------------
openai==1.56.2
langchain==0.3.7
langchain-openai==0.2.6
langchain-core==0.3.15
packaging==24.1

# -------------------------
# Interface e visualização
# -------------------------
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3


# -------------------------
# Tools
# -------------------------
langchain-community>=0.2.0
wikipedia>=1.4.0
ddgs>=5.3.0